add templated SubscriptionManager and RemoteSimulationObject family
A SubscriptionManager manages one RemoteSimulationObject Type also known as
TraCI API (e.g. PersonAPI, VehicleAPI, TrafficLightAPI, SimulationAPI, ...)

Each RemoteSimulationObject (RSO) resembles one object of the remote simulator
and should contain all possible variables the API provides. If the ID of the
RSO equals the empty string (""), the RSO will manage global  variables which
are independent of a given RSO-ID. An example of such variables are the ID_LIST,
ID_COUNT, SIM_TIME of the SimulationAPI.

Each SubscriptionManager contains a RSOFactory allowing the creation new RSO
instances with a specif set of *activated Variables*

*activated Variables* (Rationale)

Each TraCI API has access to a multitude of possible variables. However, not all
variables may be needed in all Veins/OMNeT simulation studies thus  retrieving
always all variables is not efficient. Creating new subsets of the  needed
variables for each simulation study will lead to code duplication.  Proposed
solution: Each RemoteSimulationObject has a map which indicates  which variables
are active in the current instance. The map is initialize  by a RSOFactory
instance included in the SubscriptionManager. This map also simplifies the
creation of TraCI subscription commands as well as the subsequent parsing of the
result without knowing in advanced which variables should be subscribed to.
Accessing not activated variables will throw an error.

Types of RemoteSimulationObject:

* RemoteSimulationObject: Base class for each TraCIAPI
* IMobileAgent: Interface for MobileAgents. Each RSO which resembles
  moving Agents (Vehicle, Person, Bike, ...)  should implement this interface
  to allow generic passing of mobile agents through the simulation.
* SumoVehicle: Implements IMobileAgent and resembles the VehicleAPI of SUMO
* SumoPerson: Implements IMobileAgent and resembles the PersonAPI of SUMO
* ...